sửa kiểu dữ liệu của đơn giá

+++ QuanLyKho/VIEW/fNhapHang.cs
@@ -100,7 +100,7 @@ namespace QuanLyKho.VIEW
             phieuNhap.Ma_NSX = (int)cboNSX.SelectedValue;
             phieuNhap.Ma_NV = (int)cboNhanVien.SelectedValue;
             phieuNhap.Ma_LoaiSP = (int)cboLoaiSP.SelectedValue;
-            phieuNhap.DonGia = int.Parse(txtDonGia.Text);
+            phieuNhap.DonGia = decimal.Parse(txtDonGia.Text);
             phieuNhap.SoLuong = (int)numSoLuong.Value;
             phieuNhap.TenLoai = cboLoaiSP.Text;
             phieuNhap.Ten_NV = cboNhanVien.Text.Split('-')[0];
@@ -114,9 +114,9 @@ namespace QuanLyKho.VIEW
 
         private bool ValidateData()
         {
-            int donGia;
+            decimal donGia;
             int err = 0;
-            if (!int.TryParse(txtDonGia.Text, out donGia))
+            if (!decimal.TryParse(txtDonGia.Text, out donGia))
             {
                 MessageBox.Show("Bạn chưa nhập đơn giá");
                 txtDonGia.BackColor = Color.Coral;
